requirements.txt
# MCP SDK (for Claude Desktop integration)
mcp>=0.9.0

# Google Gemini (for Gemini integration)
google-generativeai>=0.3.0
google-genai>=0.2.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
websockets>=12.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
lxml>=4.9.3

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Environment Management
python-dotenv>=1.0.0

# Rate Limiting
ratelimit>=2.2.1

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# LinkedIn API Alternative
linkedin-api>=2.2.0

# HTTP Headers
fake-useragent>=1.4.0

